web_results/lambda/requirements.txt
boto3==1.7.30
botocore==1.10.30
docutils==0.14
jmespath==0.9.3
python-dateutil==2.7.3
s3transfer==0.1.13
six==1.11.0
